+++ src/main/java/dev/maicra/eruruupatch/compat/easyfarmers/KnifeHolder.java
@@ -1,9 +1,0 @@
-package dev.maicra.eruruupatch.compat.easyfarmers;
-
-import net.minecraft.world.item.ItemStack;
-
-public interface KnifeHolder {
-    ItemStack eruruu$getKnife();
-
-    void eruruu$setKnife(ItemStack stack);
-}